+++ vm/devices/storage/disklayer_sqlite/src/lib.rs
@@ -65,14 +65,20 @@
 pub mod resolver;
 
 use anyhow::Context;
+use blocking::unblock;
 use disk_backend::DiskError;
 use disk_backend::UnmapBehavior;
 use disk_layered::LayerAttach;
 use disk_layered::LayerIo;
 use disk_layered::SectorMarker;
 use disk_layered::WriteNoOverwrite;
+use guestmem::MemoryRead;
+use guestmem::MemoryWrite;
 use inspect::Inspect;
+use parking_lot::Mutex;
+use rusqlite::Connection;
 use scsi_buffers::RequestBuffers;
+use std::sync::Arc;
 
 /// Formatting parameters provided to [`FormatOnAttachSqliteDiskLayer::new`].
 ///
@@ -91,8 +97,10 @@ pub struct IncompleteFormatParams {
 pub struct FormatParams {
     /// Should the layer be considered logically read only (i.e: a cache layer)
     pub logically_read_only: bool,
-    /// The size of the layer in bytes.
+    /// The size of the layer in bytes. Must be divisible by `sector_size`.
     pub len: u64,
+    /// The size of each sector.
+    pub sector_size: u32,
 }
 
 /// A disk layer backed by sqlite, which lazily infers its topology from the
@@ -101,15 +109,21 @@ pub struct FormatParams {
 #[non_exhaustive]
 pub struct FormatOnAttachSqliteDiskLayer {
     dbhd_path: String,
+    read_only: bool,
     format_dbhd: IncompleteFormatParams,
 }
 
 impl FormatOnAttachSqliteDiskLayer {
     /// Create a new sqlite-backed disk layer, which is formatted when it is
     /// attached.
-    pub fn new(dbhd_path: String, format_dbhd: IncompleteFormatParams) -> anyhow::Result<Self> {
+    pub fn new(
+        dbhd_path: String,
+        read_only: bool,
+        format_dbhd: IncompleteFormatParams,
+    ) -> anyhow::Result<Self> {
         Ok(Self {
             dbhd_path,
+            read_only,
             format_dbhd,
         })
     }
@@ -118,22 +132,130 @@ impl FormatOnAttachSqliteDiskLayer {
 /// A disk layer backed entirely by sqlite.
 #[derive(Inspect)]
 pub struct SqliteDiskLayer {
-    // TODO
+    #[inspect(skip)]
+    conn: Arc<Mutex<Connection>>,
+    meta: schema::DiskMeta,
+    read_only: bool,
 }
 
 impl SqliteDiskLayer {
     /// Create a new sqlite-backed disk layer.
-    pub fn new(dbhd_path: String, format_dbhd: Option<FormatParams>) -> anyhow::Result<Self> {
-        todo!()
+    pub fn new(
+        dbhd_path: String,
+        read_only: bool,
+        format_dbhd: Option<FormatParams>,
+    ) -> anyhow::Result<Self> {
+        // DEVNOTE: sqlite _really_ want to be in control of opening the file,
+        // since it also wants to read/write to the runtime "sidecar" files that
+        // get created when accessing the DB (i.e: the `*-shm` and `*-wal`
+        // files)
+        //
+        // This will make it tricky to sandbox SQLite in the future...
+        //
+        // One idea: maybe we could implement a small SQLite `vfs` shim that
+        // lets use pre-open those particular files on the caller side, and hand
+        // them to sqlite when requested (vs. having it `open()` them itself?)
+        let conn = Connection::open_with_flags(dbhd_path, {
+            use rusqlite::OpenFlags;
+
+            let mut flags = OpenFlags::SQLITE_OPEN_NO_MUTEX | OpenFlags::SQLITE_OPEN_URI;
+
+            if read_only {
+                flags |= OpenFlags::SQLITE_OPEN_READ_ONLY;
+            } else {
+                flags |= OpenFlags::SQLITE_OPEN_READ_WRITE;
+            }
+
+            // FUTURE: if/when the VFS layer is implemented, it _may_ be worth
+            // removing this flag entirely, and relying on the VFS to ensure
+            // that the (possibly blank) db file has been created. Emphasis on
+            // the word "may", as its unclear what the best approach will be
+            // until if/when we have more of the VFS infrastructure in place.
+            if format_dbhd.is_some() {
+                flags |= OpenFlags::SQLITE_OPEN_CREATE
+            }
+
+            flags
+        })?;
+
+        let meta = if let Some(FormatParams {
+            logically_read_only,
+            len,
+            sector_size,
+        }) = format_dbhd
+        {
+            use rusqlite::config::DbConfig;
+
+            // Wipe any existing contents.
+            //
+            // see https://www.sqlite.org/c3ref/c_dbconfig_defensive.html#sqlitedbconfigresetdatabase
+            conn.set_db_config(DbConfig::SQLITE_DBCONFIG_RESET_DATABASE, true)?;
+            conn.execute("VACUUM", ())?;
+            conn.set_db_config(DbConfig::SQLITE_DBCONFIG_RESET_DATABASE, false)?;
+
+            // Set core database config, and initialize table structure
+            conn.pragma_update(None, "journal_mode", "WAL")?;
+            conn.execute(schema::DEFINE_TABLE_SECTORS, [])?;
+            conn.execute(schema::DEFINE_TABLE_METADATA, [])?;
+
+            if len % sector_size as u64 != 0 {
+                anyhow::bail!(
+                    "failed to format: len={len} must be multiple of sector_size={sector_size}"
+                );
+            }
+            let sector_count = len / sector_size as u64;
+
+            let meta = schema::DiskMeta {
+                logically_read_only,
+                sector_count,
+                sector_size,
+            };
+
+            conn.execute(
+                "INSERT INTO meta VALUES (json(?))",
+                [serde_json::to_string(&meta).unwrap()],
+            )?;
+
+            meta
+        } else {
+            use rusqlite::OptionalExtension;
+            let data: String = conn
+                .query_row("SELECT json_extract(metadata, '$') FROM meta", [], |row| {
+                    row.get(0)
+                })
+                .optional()?
+                .context("missing `meta` table")?;
+            serde_json::from_str(&data)?
+        };
+
+        Ok(SqliteDiskLayer {
+            conn: Arc::new(Mutex::new(conn)),
+            meta,
+            read_only,
+        })
     }
 
-    fn write_maybe_overwrite(
+    async fn write_maybe_overwrite(
         &self,
         buffers: &RequestBuffers<'_>,
         sector: u64,
         overwrite: bool,
     ) -> Result<(), DiskError> {
-        todo!()
+        assert!(!self.read_only);
+
+        let count = buffers.len() / self.meta.sector_size as usize;
+        tracing::trace!(sector, count, "write");
+
+        let buf = buffers.reader().read_all()?;
+        unblock({
+            let conn = self.conn.clone();
+            let sector_size = self.meta.sector_size;
+            move || write_sectors(conn, sector_size, sector, buf, overwrite)
+        })
+        .await
+        .map_err(|e| DiskError::Io(std::io::Error::new(std::io::ErrorKind::Other, e)))?;
+
+        Ok(())
     }
 }
 
@@ -146,18 +268,24 @@ impl LayerAttach for FormatOnAttachSqliteDiskLayer {
         lower_layer_metadata: Option<disk_layered::DiskLayerMetadata>,
     ) -> Result<Self::Layer, Self::Error> {
         let len = {
-            let lower_len = lower_layer_metadata.map(|m| m.sector_count * m.sector_size as u64);
+            let lower_len = lower_layer_metadata
+                .as_ref()
+                .map(|m| m.sector_count * m.sector_size as u64);
             self.format_dbhd
                 .len
                 .or(lower_len)
                 .context("no base layer to infer sector_count from")?
         };
+        // FUTURE: make sector-size configurable
+        let sector_size = lower_layer_metadata.map(|x| x.sector_size).unwrap_or(512);
 
         SqliteDiskLayer::new(
             self.dbhd_path,
+            self.read_only,
             Some(FormatParams {
                 logically_read_only: self.format_dbhd.logically_read_only,
                 len,
+                sector_size,
             }),
         )
     }
@@ -169,27 +297,27 @@ impl LayerIo for SqliteDiskLayer {
     }
 
     fn sector_count(&self) -> u64 {
-        todo!()
+        self.meta.sector_count
     }
 
     fn sector_size(&self) -> u32 {
-        todo!()
+        self.meta.sector_size
     }
 
     fn is_read_only(&self) -> bool {
-        todo!()
+        self.read_only
     }
 
     fn disk_id(&self) -> Option<[u8; 16]> {
-        todo!()
+        None
     }
 
     fn physical_sector_size(&self) -> u32 {
-        todo!()
+        self.meta.sector_size
     }
 
     fn is_fua_respected(&self) -> bool {
-        todo!()
+        true
     }
 
     async fn read(
@@ -198,7 +326,36 @@ impl LayerIo for SqliteDiskLayer {
         sector: u64,
         mut marker: SectorMarker<'_>,
     ) -> Result<(), DiskError> {
-        todo!()
+        let sector_count = (buffers.len() / self.meta.sector_size as usize) as u64;
+        let end_sector = sector + sector_count;
+        tracing::trace!(sector, sector_count, "read");
+        if end_sector > self.meta.sector_count {
+            return Err(DiskError::IllegalBlock);
+        }
+
+        let valid_sectors = unblock({
+            let conn = self.conn.clone();
+            let end_sector = sector + (buffers.len() as u64 / self.meta.sector_size as u64);
+            let sector_size = self.meta.sector_size;
+            move || read_sectors(conn, sector_size, sector, end_sector)
+        })
+        .await
+        .map_err(|e| DiskError::Io(std::io::Error::new(std::io::ErrorKind::Other, e)))?;
+
+        for (s, data) in valid_sectors {
+            let offset = (s - sector) as usize * self.meta.sector_size as usize;
+            let subrange = buffers.subrange(offset, self.meta.sector_size as usize);
+            let mut writer = subrange.writer();
+            match data {
+                SectorKind::AllZero => writer.zero(self.meta.sector_size as usize)?,
+                SectorKind::AllOne => writer.fill(0xff, self.meta.sector_size as usize)?,
+                SectorKind::Data(data) => writer.write(&data)?,
+            };
+
+            marker.set(s);
+        }
+
+        Ok(())
     }
 
     async fn write(
@@ -207,7 +364,7 @@ impl LayerIo for SqliteDiskLayer {
         sector: u64,
         _fua: bool,
     ) -> Result<(), DiskError> {
-        self.write_maybe_overwrite(buffers, sector, true)
+        self.write_maybe_overwrite(buffers, sector, true).await
     }
 
     fn write_no_overwrite(&self) -> Option<impl WriteNoOverwrite> {
@@ -215,7 +372,13 @@ impl LayerIo for SqliteDiskLayer {
     }
 
     async fn sync_cache(&self) -> Result<(), DiskError> {
-        todo!()
+        tracing::trace!("sync_cache");
+
+        (self.conn.lock())
+            .pragma_update(None, "wal_checkpoint", "FULL")
+            .map_err(|e| DiskError::Io(std::io::Error::new(std::io::ErrorKind::Other, e)))?;
+
+        Ok(())
     }
 
     async fn unmap(
@@ -225,7 +388,19 @@ impl LayerIo for SqliteDiskLayer {
         _block_level_only: bool,
         next_is_zero: bool,
     ) -> Result<(), DiskError> {
-        todo!()
+        tracing::trace!(sector_offset, sector_count, "unmap");
+        if sector_offset + sector_count > self.meta.sector_count {
+            return Err(DiskError::IllegalBlock);
+        }
+
+        unblock({
+            let conn = self.conn.clone();
+            move || unmap_sectors(conn, sector_offset, sector_count, next_is_zero)
+        })
+        .await
+        .map_err(|e| DiskError::Io(std::io::Error::new(std::io::ErrorKind::Other, e)))?;
+
+        Ok(())
     }
 
     fn unmap_behavior(&self) -> UnmapBehavior {
@@ -245,6 +420,198 @@ impl WriteNoOverwrite for SqliteDiskLayer {
         buffers: &RequestBuffers<'_>,
         sector: u64,
     ) -> Result<(), DiskError> {
-        self.write_maybe_overwrite(buffers, sector, false)
+        self.write_maybe_overwrite(buffers, sector, false).await
+    }
+}
+
+#[allow(clippy::large_enum_variant)]
+enum SectorKind {
+    AllZero,
+    AllOne,
+    Data(Vec<u8>),
+}
+
+// FUTURE: pass RequestBuffers into this function directly, and avoid the alloc
+// overhead.
+fn read_sectors(
+    conn: Arc<Mutex<Connection>>,
+    sector_size: u32,
+    start_sector: u64,
+    end_sector: u64,
+) -> Result<Vec<(u64, SectorKind)>, rusqlite::Error> {
+    let conn = conn.lock();
+
+    let mut select_stmt = conn.prepare_cached(
+        "SELECT sector, data
+        FROM sectors
+        WHERE sector >= ? AND sector < ?
+        ORDER BY sector ASC",
+    )?;
+    let mut rows = select_stmt.query(rusqlite::params![start_sector, end_sector])?;
+
+    // reuse the same blob handle across all reads.
+    //
+    // this handle isn't easy to persist across multiple calls to
+    // `read_sectors`, due to the lifetime restrictions imposed by using a
+    // `Arc<Mutex<Connection>>`.
+    let Some(first_row) = rows.next()? else {
+        return Ok(Vec::new());
+    };
+    let mut sector: u64 = first_row.get(0)?;
+    let mut blob = conn.blob_open(
+        rusqlite::DatabaseName::Main,
+        "sectors",
+        "data",
+        sector as i64,
+        true,
+    )?;
+
+    let mut buf = vec![0; sector_size as usize];
+    let mut res = Vec::new();
+    loop {
+        let data = match blob.len() {
+            0 => SectorKind::AllZero,
+            1 => SectorKind::AllOne,
+            len => {
+                if len != sector_size as usize {
+                    return Err(rusqlite::Error::BlobSizeError);
+                }
+                blob.read_at(&mut buf, 0)?;
+                SectorKind::Data(buf.clone())
+            }
+        };
+
+        res.push((sector, data));
+
+        let Some(next_row) = rows.next()? else {
+            break;
+        };
+
+        sector = next_row.get(0)?;
+        blob.reopen(sector as i64)?;
+    }
+
+    Ok(res)
+}
+
+// FUTURE: pass RequestBuffers into this function directly, and avoid the alloc
+// overhead.
+fn write_sectors(
+    conn: Arc<Mutex<Connection>>,
+    sector_size: u32,
+    mut sector: u64,
+    buf: Vec<u8>,
+    overwrite: bool,
+) -> Result<(), rusqlite::Error> {
+    let mut conn = conn.lock();
+
+    let tx = conn.transaction()?;
+    {
+        let mut stmt = if overwrite {
+            tx.prepare_cached("INSERT OR REPLACE INTO sectors (sector, data) VALUES (?, ?)")?
+        } else {
+            tx.prepare_cached("INSERT OR IGNORE INTO sectors (sector, data) VALUES (?, ?)")?
+        };
+
+        let chunks = buf.chunks_exact(sector_size as usize);
+        if !chunks.remainder().is_empty() {
+            return Err(rusqlite::Error::BlobSizeError);
+        }
+        for chunk in chunks {
+            let chunk = if chunk.iter().all(|x| *x == 0) {
+                &[]
+            } else if chunk.iter().all(|x| *x == 0xff) {
+                &[0]
+            } else {
+                chunk
+            };
+
+            stmt.execute(rusqlite::params![sector, chunk])?;
+            sector += 1;
+        }
+    }
+    tx.commit()?;
+
+    Ok(())
+}
+
+fn unmap_sectors(
+    conn: Arc<Mutex<Connection>>,
+    sector_offset: u64,
+    sector_count: u64,
+    next_is_zero: bool,
+) -> Result<(), rusqlite::Error> {
+    let mut conn = conn.lock();
+
+    if next_is_zero {
+        let mut clear_stmt =
+            conn.prepare_cached("DELETE FROM sectors WHERE sector BETWEEN ? AND ?")?;
+        clear_stmt.execute(rusqlite::params![
+            sector_offset,
+            sector_offset + sector_count - 1
+        ])?;
+    } else {
+        let tx = conn.transaction()?;
+        {
+            let mut stmt =
+                tx.prepare_cached("INSERT OR REPLACE INTO sectors (sector, data) VALUES (?, ?)")?;
+
+            for sector in sector_offset..(sector_offset + sector_count) {
+                let zero_blob = &[];
+                stmt.execute(rusqlite::params![sector, zero_blob])?;
+            }
+        }
+        tx.commit()?;
+    }
+
+    Ok(())
+}
+
+mod schema {
+    use inspect::Inspect;
+    use serde::Deserialize;
+    use serde::Serialize;
+
+    // DENOTE: SQLite actually saves the _plaintext_ of CREATE TABLE
+    // statements in its file format, which makes it a pretty good place to
+    // stash inline comments about the schema being used
+    //
+    // DEVNOTE: the choice to use the len of the blob as a marker for all
+    // zero / all one sectors has not been profiled relative to other
+    // implementation (e.g: having a third "kind" column).
+    pub const DEFINE_TABLE_SECTORS: &str = r#"
+CREATE TABLE sectors (
+    -- schema includes a minimal "fast path" for skipping all-zero
+    -- and all-one sectors.
+    --
+    -- if len == 0: represents all 0x00 sector
+    -- if len == 1: represents all 0xff sector
+    --
+    -- otherwise, data has len == SECTOR_SIZE, and contains the raw
+    -- sector data.
+    sector INTEGER NOT NULL,
+    data   BLOB NOT NULL,
+    PRIMARY KEY (sector)
+)
+"#; // TODO?: enforce sqlite >3.37.0 so we can use STRICT
+
+    // DEVNOTE: Given that this is a singleton table, we might as well use JSON
+    // + serde to store whatever metadata we want here, vs. trying to bend our
+    // metadata structure to sqlite's native data types.
+    //
+    // Using JSON (vs, say, protobuf) has the added benefit of allowing existing
+    // external sqlite tooling to more easily read and manipulate the metadata
+    // using sqlite's built-in JSON handling functions.
+    pub const DEFINE_TABLE_METADATA: &str = r#"
+CREATE TABLE meta (
+    metadata TEXT NOT NULL -- stored as JSON
+)
+"#;
+
+    #[derive(Debug, PartialEq, PartialOrd, Eq, Ord, Serialize, Deserialize, Inspect)]
+    pub struct DiskMeta {
+        pub logically_read_only: bool,
+        pub sector_count: u64,
+        pub sector_size: u32,
     }
 }
